Update get_max_spanning_tree.py
This problem seems to require us to find the maximum spanning tree of a d-dimensional lattice. We are given d (the number of dimensions) and k (a constant), and then we are given d integers describing the size of each dimension. We need to find the maximum spanning tree of this lattice and return the result modulo 10^9+7.

To solve this problem, we can use Kruskal's algorithm to find the maximum spanning tree of the lattice. Kruskal's algorithm starts with all edges in the graph and repeatedly removes the edge with the smallest weight until there are no more edges or the remaining edges form a tree. We can modify Kruskal's algorithm to work with a d-dimensional lattice by treating each point in the lattice as a node and each edge between adjacent points as an edge in the graph. We can compute the weight of each edge as the Euclidean distance between the two points it connects. Then we can run Kruskal's algorithm on this graph to find the maximum spanning tree.

To compute the Euclidean distance between two points in a d-dimensional lattice, we can use the formula:

distance = sqrt((x1 - x2)^2 + (y1 - y2)^2 + ... + (d1 - d2)^2)

where (x1, y1, ..., d1) and (x2, y2, ..., d2) are the coordinates of the two points.

Finally, we need to return the result modulo 10^9+7. This is a common technique in competitive programming to avoid integer overflow when working with large numbers.