Rename html_parser.py to python/html_parser.py
The code defines a class named MyHTMLParser that inherits from the HTMLParser class from the html.parser module. This class is used to parse HTML data and extract information from it. The class has several methods that can be overridden to handle different types of HTML tags and data:

handle_starttag(self, tag, attrs): This method is called whenever a start tag is encountered in the HTML data. It takes two arguments: tag, which is the name of the tag, and attrs, which is a list of tuples containing the attributes and their values for the tag. This method can be overridden to perform operations when a specific tag is encountered.
handle_endtag(self, tag): This method is called whenever an end tag is encountered in the HTML data. It takes one argument: tag, which is the name of the tag. This method can be overridden to perform operations when a specific tag is closed.
handle_data(self, data): This method is called whenever text data is encountered in the HTML data. It takes one argument: data, which is the text data. This method can be overridden to perform operations on the text data.
handle_error(self, message): This method is called whenever an error is encountered while parsing the HTML data. It takes one argument: message, which is a string describing the error. This method can be overridden to handle errors in a specific way.
The MyHTMLParser class also defines two additional methods: parse_html(self, html_string) and parse_url(self, url). These methods can be used to parse HTML data from a string or a URL, respectively.

parse_html(self, html_string): This method takes one argument: html_string, which is the HTML data to be parsed. This method calls the feed method of the HTMLParser class to parse the HTML data.
parse_url(self, url): This method takes one argument: url, which is the URL of the HTML data to be parsed. This method uses the urllib.request.urlopen method to open the URL and read the HTML data. It then calls the parse_html method to parse the HTML data.
The MyHTMLParser class also defines an init method that calls the reset method of the HTMLParser class to reset the parser’s internal state. It also defines an errors attribute that is a list of error messages encountered while parsing the HTML data.

The code also includes an example of how to use the MyHTMLParser class. In the example, an instance of the MyHTMLParser class is created and its parse_url method is called with a URL as an argument. The parse_url method opens the URL, reads the HTML data, and parses it. The handle_starttag, handle_endtag, and handle_data methods of the MyHTMLParser class are overridden to print information about the HTML data as it is parsed.

The code demonstrates how to use the HTMLParser class to extract information from HTML data.